UI以外の部分をModelに分離
