+++ Assets/MeshDeleterWithTexture/Editor/Models/MeshDeleterWithTextureModel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a1042ff14fbb1d54983b1b4b05622b8e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
